backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
firebase-admin==6.5.0
pydantic==2.8.2
pytest==8.3.2
httpx==0.27.2
python-dotenv==1.0.1
